Check list/read methods return CheckResponse instead of Check
Since the returned value from the pingdom API is not always a
complete check it should be more clear to return an explictly
different type of object.

This will also make it easier to support polymorphic Check types
as input to Create/Update since the Read/List responses won't differ.