Extended NewClient to allow Multi-User auth
We want to be able to authenticate in multi-user Pingdom
accounts. For this we need to capture an `AccountEmail`
variable, which is the email address of the account holder.
A NewMultiUserClient constructor has been created, which
extends the existing NewClient constructor and allows the
AccountEmail variable to be passed. This will be passed
in the `Account-Email` header when making API calls.