Add HTTP check specific details to API responses
Change the HTTP API responses to support the specific details of HTTP
Checks, as described in [1].

We add the fields:
 * `CheckResponse.Type.Name`, check type name,
 * `CheckResponse.Type.HTTP`,  with the details of a HTTP check or nil 
     if it is other type.

Note that the response for list and get one check differ, as list returns 
`type` as a string with the check type ("http", "ping"...) and get details
 returns a hash with all the specific details. 

To support that case, we add a field `Type` with go type 
`CheckResponseType` which implements `json.Unmarshaller` interface [2].
In the Unmarshall function we check if the json is a plain string or a hash, 
continuing the unmarshalling at that point in the second case.

This approach would allow extend the library with details of other check 
types details without refactoring the signature of the types, by adding
additional fields to `CheckResponseType`.


[1] https://www.pingdom.com/resources/api#MethodGet+Detailed+Check+Information
[2] https://golang.org/pkg/encoding/json/#Unmarshaler