pingdom: parse team ids in check instance response
There are three issues with the current parsing of teams as part of
a check instance response:

- the client must provide `?include_teams=true` to receive teams in
the API response
- the team response has the key "teams" not "teamids"
- the team response is a list of structs with "name" and "id" fields,
not a list of integers.

Unfortunately Pingdom returns "id" in a GET to the /teams list as
a string, but returns it as an integer in a GET to /checks, so we
create a separate struct for team responses as part of the check
response. Update the tests to handle this case.

We could add an optional parameter to ask callers whether they'd like
to include teams as part of the response, but it seems cheap enough to
always include them, and adding an option in the function interface
would probably be a breaking API change.

This work was sponsored by [Notion](https://www.notion.so).

Updates russellcardullo/terraform-provider-pingdom#26.