Add acceptance tests
Adds tests that will actually call the Pingdom API to validate
the library.  The tests will attempt to be self-cleaning as
much as possible.

The acceptance tests are opt-in.  Develoeprs can enable those
by calling the appropriate make target, or outside of make
by setting an environment variable `PINGDOM_ACCEPTANCE=1`.