Unmarshal check severity_level
The Pingdom API allows an optional parameter when getting checks: .
This returns a 'severity_level' field for each check: 'HIGH' or 'LOW'.

This can be changed in the Pingdom UI using a radio button: 'Check Importance' which
defaults to 'High Severity'. There doesn't seem to be an equivalent parameter for check creation.